chore: Unify import symbol acquisition.
All identifier acquisition for imports ("derive", etc.) are all accessed from ImportRequirements, generating a new identifier, and requiring that import to be available. This is a prerequisite for handling other forms of authored imports, and cleans up import management.

* Allow some functions to return optionals (createDeriveCallForExpression) rather than relying on consumer to check equality of input statement
* Aligns some typing between TransformationContext and EmitterContext
* Remove helper accumulation in emitters (handled when requesting a symbol)
* Remove import-resolver, extraneous/unused functions in builtins.ts; split up the builtin AST helpers
* To disambiguate from our own TransformationContext, ts.TransformationContext variable usage renamed to tsContext
