+++ packages/ts-transformers/src/transformers/schema-injection.ts
@@ -104,18 +104,25 @@ function collectFunctionSchemaTypeNodes(
 }
 
 function createToSchemaCall(
-  factory: ts.NodeFactory,
+  { imports, factory, sourceFile }: Pick<
+    TransformationContext,
+    "imports" | "factory" | "sourceFile"
+  >,
   typeNode: ts.TypeNode,
 ): ts.CallExpression {
+  const identifier = imports.getIdentifier({ factory, sourceFile }, {
+    module: "commontools",
+    name: "toSchema",
+  });
   return factory.createCallExpression(
-    factory.createIdentifier("toSchema"),
+    identifier,
     [typeNode],
     [],
   );
 }
 
 function prependSchemaArguments(
-  factory: ts.NodeFactory,
+  context: Pick<TransformationContext, "factory" | "imports" | "sourceFile">,
   node: ts.CallExpression,
   argumentTypeNode: ts.TypeNode,
   argumentType: ts.Type | undefined,
@@ -124,8 +131,14 @@ function prependSchemaArguments(
   typeRegistry?: TypeRegistry,
   checker?: ts.TypeChecker,
 ): ts.CallExpression {
-  const argSchemaCall = createToSchemaCall(factory, argumentTypeNode);
-  const resSchemaCall = createToSchemaCall(factory, resultTypeNode);
+  const argSchemaCall = createToSchemaCall(
+    context,
+    argumentTypeNode,
+  );
+  const resSchemaCall = createToSchemaCall(
+    context,
+    resultTypeNode,
+  );
 
   // Register Types if they were inferred (not from original source)
   if (typeRegistry && checker) {
@@ -137,7 +150,7 @@ function prependSchemaArguments(
     }
   }
 
-  return factory.createCallExpression(
+  return context.factory.createCallExpression(
     node.expression,
     undefined,
     [argSchemaCall, resSchemaCall, ...node.arguments],
@@ -149,18 +162,8 @@ export class SchemaInjectionTransformer extends Transformer {
     return hasCtsEnableDirective(context.sourceFile);
   }
   transform(context: TransformationContext): ts.SourceFile {
-    const { sourceFile, transformation, checker } = context;
+    const { sourceFile, tsContext: transformation, checker, imports } = context;
     const typeRegistry = context.options.typeRegistry;
-
-    let requestedToSchema = false;
-    const ensureToSchemaImport = (): void => {
-      if (requestedToSchema) return;
-      requestedToSchema = true;
-      context.imports.require({
-        module: "commontools",
-        name: "toSchema",
-      });
-    };
 
     const visit = (node: ts.Node): ts.Node => {
       if (!ts.isCallExpression(node)) {
@@ -175,13 +178,7 @@ export class SchemaInjectionTransformer extends Transformer {
           const factory = transformation.factory;
           const schemaArgs = typeArgs.map((typeArg) => typeArg).map((
             typeArg,
-          ) =>
-            factory.createCallExpression(
-              factory.createIdentifier("toSchema"),
-              [typeArg],
-              [],
-            )
-          );
+          ) => createToSchemaCall(context, typeArg));
 
           const argsArray = Array.from(node.arguments);
           let remainingArgs = argsArray;
@@ -191,8 +188,6 @@ export class SchemaInjectionTransformer extends Transformer {
           ) {
             remainingArgs = argsArray.slice(1);
           }
-
-          ensureToSchemaImport();
 
           const updated = factory.createCallExpression(
             node.expression,
@@ -215,18 +210,8 @@ export class SchemaInjectionTransformer extends Transformer {
           if (!eventType || !stateType) {
             return ts.visitEachChild(node, visit, transformation);
           }
-          const toSchemaEvent = factory.createCallExpression(
-            factory.createIdentifier("toSchema"),
-            [eventType],
-            [],
-          );
-          const toSchemaState = factory.createCallExpression(
-            factory.createIdentifier("toSchema"),
-            [stateType],
-            [],
-          );
-
-          ensureToSchemaImport();
+          const toSchemaEvent = createToSchemaCall(context, eventType);
+          const toSchemaState = createToSchemaCall(context, stateType);
 
           const updated = factory.createCallExpression(
             node.expression,
@@ -254,18 +239,8 @@ export class SchemaInjectionTransformer extends Transformer {
                 factory.createKeywordTypeNode(ts.SyntaxKind.UnknownKeyword);
 
               if (eventParam || stateParam) {
-                const toSchemaEvent = factory.createCallExpression(
-                  factory.createIdentifier("toSchema"),
-                  [eventType],
-                  [],
-                );
-                const toSchemaState = factory.createCallExpression(
-                  factory.createIdentifier("toSchema"),
-                  [stateType],
-                  [],
-                );
-
-                ensureToSchemaImport();
+                const toSchemaEvent = createToSchemaCall(context, eventType);
+                const toSchemaState = createToSchemaCall(context, stateType);
 
                 const updated = factory.createCallExpression(
                   node.expression,
@@ -288,9 +263,8 @@ export class SchemaInjectionTransformer extends Transformer {
           resultType: ts.TypeNode,
           resultTypeValue: ts.Type | undefined,
         ): ts.Node => {
-          ensureToSchemaImport();
           const updated = prependSchemaArguments(
-            factory,
+            context,
             node,
             argumentType,
             argumentTypeValue,
@@ -357,9 +331,8 @@ export class SchemaInjectionTransformer extends Transformer {
           resultType: ts.TypeNode,
           resultTypeValue: ts.Type | undefined,
         ): ts.Node => {
-          ensureToSchemaImport();
           const updated = prependSchemaArguments(
-            factory,
+            context,
             node,
             argumentType,
             argumentTypeValue,
